Move AAB code into its own files
