+++ simulant.files
@@ -700,3 +700,5 @@ simulant/math/mat3.h
 simulant/math/mat3.cpp
 simulant/math/mat4.h
 simulant/math/mat4.cpp
+simulant/math/aabb.h
+simulant/math/aabb.cpp
